requirements.txt
azure-functions
fastmcp>=3.0
azure-data-tables
msgraph-sdk>=1.56.0
azure-identity>=1.25.3
